refactor tests - wip
